refactor(auth): move forgot password mutation to hooks directory
The mutation was moved from the api directory to the hooks directory to better organize auth-related hooks and follow the project's structure conventions.
